♻️ : 여러 모듈에서 데이터 처리 및 상태 결정 로직 개선, null 체크 추가 및 코드 정리

+++ src/modules/account/application/account-application.command.service.ts
@@ -218,7 +218,11 @@ export class AccountApplicationCommandService {
 
       const data =
         await this.rankCoreService.aggregateRankStatsByUserId(userId);
-      if (data[prevTeamId] && data[prevTeamId].getTotalCount() === 0) {
+      if (
+        prevTeamId !== value &&
+        data[prevTeamId] &&
+        data[prevTeamId].getTotalCount() === 0
+      ) {
         await this.rankCoreService.deleteRank(prevTeamId, userId);
         await this.rankingRedisService.deleteRankByUserId(userId, [prevTeamId]);
         delete data[prevTeamId];
